+++ row.go
@@ -4,7 +4,6 @@ import (
 	"database/sql"
 	"errors"
 	"reflect"
-	"time"
 )
 
 var (
@@ -108,33 +107,4 @@ func (r *Row) Bind(dest any) error {
 
 	// Make sure the query can be processed to completion with no errors.
 	return r.rows.Close()
-}
-
-func scanTo(dest any, destValue reflect.Value, cols []string, rows *sql.Rows) (bool, error) {
-	var err error
-	switch b := dest.(type) {
-	case *int, *int8, *int16, *int32, *int64,
-		*uint, *uint8, *uint16, *uint32, *uint64, *[]byte,
-		*uintptr, *float32, *float64, *bool, *string, *time.Time,
-		sql.Scanner:
-		err = rows.Scan(dest)
-		if err != nil {
-			return true, err
-		}
-
-		return true, rows.Close()
-	case Binder:
-		err = rows.Scan(b.Bind(destValue, cols)...)
-		if err != nil {
-			return true, err
-		}
-		return true, rows.Close()
-	}
-
-	return false, nil
-}
-
-func scanToStruct(v reflect.Value, cols []string, rows *sql.Rows) error {
-	b := getStructBinder(v.Type(), v)
-	return rows.Scan(b.Bind(v, cols)...)
 }