+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - added `BuilderOption` on `SetMap` on `InsertBuilder` and `UpdateBuilder` (#1)
 - added `WithAllow` BuilderOption (#2)
-
+- added custom `Binder` support on `Bind` (#4)
 ### Fixed
-- fixed `sql.Scanner` support on `bind` (#2)
+- fixed `sql.Scanner` support on `Bind` (#2)
   
 ## [1.0.0] - 2024-01-31